upload fixbug

+++ operation/views.py
@@ -152,46 +152,47 @@ def upload_upload(request):
     if int(flag) == 0:
         #开始传输
         # orm = upload_files.objects.get(file_name=file_name)
-        cmdLine = []
-        cmdLine.append('rsync')
-        cmdLine.append('--progress')
-        cmdLine.append('uploads/%s' % file_name)
-        cmdLine.append('%s:%s' % (rsync_ip,rsync_dir))
-        tmpFile = BASE_DIR + "/tmp/upload.tmp"  #临时生成一个文件
-        fpWrite = open(tmpFile,'w')
-        with open(BASE_DIR + '/tmp/rsync_status_file.tmp','w') as f:
-            pass
-        process = subprocess.Popen(cmdLine,stdout = fpWrite,stderr = subprocess.PIPE);
-        while True:
-            fpRead = open(tmpFile,'r')  #这里又重新创建了一个文件读取对象，不知为何，用上面的就是读不出来，改w+也不>行
-            lines = fpRead.readlines()
-            for line in lines:
-                a = re.search(r'\d+%',line)
-                if a:
-                    with open(BASE_DIR + '/tmp/percent.tmp','a') as f:
-                            f.write(a.group())
-            if  process.poll() == 0:
-                break;
-            elif process.poll() == None:
+        rsync_dest = request.POST.get('rsync_dest')
+        if rsync_dest:
+            rsync_ip = CENTER_SERVER[rsync_dest][0]
+            rsync_dir = CENTER_SERVER[rsync_dest][2]
+            cmdLine = []
+            cmdLine.append('rsync')
+            cmdLine.append('--progress')
+            cmdLine.append('uploads/%s' % file_name)
+            cmdLine.append('%s:%s' % (rsync_ip,rsync_dir))
+            tmpFile = BASE_DIR + "/tmp/upload.tmp"  #临时生成一个文件
+            fpWrite = open(tmpFile,'w')
+            with open(BASE_DIR + '/tmp/rsync_status_file.tmp','w') as f:
                 pass
-            else:
-                break
-            fpWrite.truncate()  #此处清空文件，等待记录下一次输出的进度
-            fpRead.close()
-            time.sleep(0.7)
-        fpWrite.close()
-        os.remove(BASE_DIR + '/tmp/rsync_status_file.tmp')
-    #    error = process.stderr.read()
-    #    if not error == None:
-    #        print 'error info:%s' % error
-        os.remove(tmpFile) #删除临时文件
-        os.remove(BASE_DIR + '/tmp/percent.tmp')
-        return HttpResponse(simplejson.dumps({'code':0,'msg':u'文件传输成功'}),content_type="application/json")
+            process = subprocess.Popen(cmdLine,stdout = fpWrite,stderr = subprocess.PIPE);
+            while True:
+                fpRead = open(tmpFile,'r')  #这里又重新创建了一个文件读取对象，不知为何，用上面的就是读不出来，改w+也不>行
+                lines = fpRead.readlines()
+                for line in lines:
+                    a = re.search(r'\d+%',line)
+                    if a:
+                        with open(BASE_DIR + '/tmp/percent.tmp','a') as f:
+                                f.write(a.group())
+                if  process.poll() == 0:
+                    break;
+                elif process.poll() == None:
+                    pass
+                else:
+                    break
+                fpWrite.truncate()  #此处清空文件，等待记录下一次输出的进度
+                fpRead.close()
+                time.sleep(0.7)
+            fpWrite.close()
+            os.remove(BASE_DIR + '/tmp/rsync_status_file.tmp')
+        #    error = process.stderr.read()
+        #    if not error == None:
+        #        print 'error info:%s' % error
+            os.remove(tmpFile) #删除临时文件
+            os.remove(BASE_DIR + '/tmp/percent.tmp')
+            return HttpResponse(simplejson.dumps({'code':0,'msg':u'文件传输成功'}),content_type="application/json")
     elif int(flag) == 1:
         #获取百分比
-        rsync_dest = request.POST.get('rsync_dest')
-        rsync_ip = CENTER_SERVER[rsync_dest][0]
-        rsync_dir = CENTER_SERVER[rsync_dest][2]
         if os.path.exists(BASE_DIR + '/tmp/rsync_status_file.tmp'):
             process = 1
         else:
